revert(nav): restore 4-slot bottom nav + shared Dashboard landing
Reverts persona-aware navigation introduced in #57-#67 and follow-ups.
The persona-shaped nav (5 slots per role) was not functional in practice:
directors lost their Dashboard, the bar reshuffled on every context
switch, and the Más sheet workflow was confusing. Restore the prior
4-slot static nav (Dashboard | Clases | Actividades | Perfil) and route
every authenticated user to the shared Dashboard after login.

Changes:
- router.dart: drop personaLandingRoute/resolvePersona; redirect logic
  lands every user on RouteNames.homeDashboard. Replace
  _CoordinatorShell + coordinator StatefulShellRoute with plain
  GoRoutes for /coordinator/* destinations (still reachable via
  push/go). Restore _NavItemConfig / _navItemsConfig / _filterNavItems
  and 4-slot _MainShell.
- Delete lib/core/persona/ and lib/core/nav/destinations.dart.
- Restore lib/features/dashboard/presentation/widgets/quick_access_grid.dart
  from pre-shared-destinations version (b1baae6^).
- dashboard_view: re-add QuickAccessGrid between QuickStatsCard and
  UpcomingActivitiesCard; drop showMoreSheet wiring.
- profile_view: drop onMoreTap entry point and the "Más opciones" tile.
- Delete persona/coordinator/landing/nav-drift tests +
  persona-nav integration sample. Update coordinator_clubs_list_view_test
  to drop persona resolver assertions.

All 225 unit tests pass; flutter analyze clean.
