--- lib/core/persona/index.dart
@@ -1,16 +1,0 @@
-/// Persona module — barrel export.
-///
-/// Import this file to access all public types from the persona module:
-///
-/// ```dart
-/// import 'package:sacdia_app/core/persona/index.dart';
-/// ```
-library;
-
-export 'nav_slot.dart';
-export 'persona.dart';
-export 'persona_nav_config.dart';
-export 'persona_providers.dart';
-export 'persona_resolver.dart';
-export 'widgets/nav_badge.dart';
-export 'widgets/more_sheet.dart';
