+++ lib/features/dashboard/presentation/views/dashboard_view.dart
@@ -18,11 +18,11 @@ import '../providers/dashboard_providers.dart';
 import '../widgets/club_info_card.dart';
 import '../widgets/current_class_card.dart';
 import '../widgets/quick_stats_card.dart';
+import '../widgets/quick_access_grid.dart';
 import '../widgets/upcoming_activities_card.dart';
 import '../widgets/membership_status_banner.dart';
 import '../widgets/welcome_header.dart';
 import '../../../enrollment/presentation/widgets/enrollment_status_card.dart';
-import 'package:sacdia_app/core/persona/widgets/more_sheet.dart';
 
 /// Vista principal del dashboard - Estilo "Scout Vibrante"
 class DashboardView extends ConsumerWidget {
@@ -88,9 +88,6 @@ class DashboardView extends ConsumerWidget {
                         userAvatar: dashboard.userAvatar ?? authAvatar,
                         onNotificationsTap: () =>
                             context.push(RouteNames.notificationsInbox),
-                        // T-27: Más sheet entry point #1 — action bar button
-                        onMoreTap: () =>
-                            showMoreSheet(context: context, ref: ref),
                       ),
                     ),
 
@@ -138,6 +135,10 @@ class DashboardView extends ConsumerWidget {
                             honorsCompleted: dashboard.honorsCompleted,
                             honorsInProgress: dashboard.honorsInProgress,
                           ),
+                          const SizedBox(height: 16),
+
+                          // Acceso rápido — launcher de módulos
+                          const QuickAccessGrid(),
                           const SizedBox(height: 16),
 
                           // Upcoming activities
