+++ test/features/coordinator/presentation/views/coordinator_clubs_list_view_test.dart
@@ -9,13 +9,10 @@
 /// 6. Router: coordinatorClubs route constant matches expected path.
 library;
 
-import 'dart:async';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sacdia_app/core/config/route_names.dart';
-import 'package:sacdia_app/core/persona/index.dart';
 import 'package:sacdia_app/features/auth/domain/entities/authorization_snapshot.dart';
 import 'package:sacdia_app/features/auth/domain/entities/user_entity.dart';
 import 'package:sacdia_app/features/auth/presentation/providers/auth_providers.dart';
@@ -107,13 +104,7 @@ Future<void> _pumpView(
 void main() {
   // ── Unit: RBAC gating ─────────────────────────────────────────────────────
 
-  group('RBAC: coordinator persona resolves correctly', () {
-    test('coordinator role maps to Persona.coordinador', () {
-      final user = _coordinatorUser();
-      final persona = resolvePersona(user.authorization);
-      expect(persona, Persona.coordinador);
-    });
-
+  group('RBAC: coordinator route wiring', () {
     test('coordinatorClubs route is /coordinator/clubs', () {
       expect(RouteNames.coordinatorClubs, '/coordinator/clubs');
     });
